Relocate InMemoryCompactionJobStatusStore and test file
