+++ java/clients/src/test/java/sleeper/clients/admin/testutils/AdminClientStatusStoreHolderTest.java
@@ -18,10 +18,10 @@ package sleeper.clients.admin.testutils;
 import org.junit.jupiter.api.Test;
 
 import sleeper.compaction.core.task.CompactionTaskStatusStore;
-import sleeper.compaction.core.testutils.InMemoryCompactionJobStatusStore;
 import sleeper.compaction.core.testutils.InMemoryCompactionTaskStatusStore;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.tracker.compaction.job.CompactionJobTracker;
+import sleeper.core.tracker.compaction.job.InMemoryCompactionJobStatusStore;
 import sleeper.ingest.core.job.status.IngestJobStatusStore;
 import sleeper.ingest.core.task.IngestTaskStatusStore;
